refactor: simplify AnalysisFormBase
